survey and favorites pages are working; weather still needs implementing and most pages need styling

+++ module-3/12-Capstone/java/src/main/java/com/techelevator/npgeek/model/Park.java
@@ -19,8 +19,7 @@ public class Park {
 	private String parkDescription;
 	private long entryFee;
 	private long numberOfAnimalSpecies;
-	private int numberOfSurveys;
-	
+	private long numberOfSurveys;
 	public String getParkCode() {
 		return parkCode;
 	}
@@ -111,12 +110,13 @@ public class Park {
 	public void setNumberOfAnimalSpecies(long numberOfAnimalSpecies) {
 		this.numberOfAnimalSpecies = numberOfAnimalSpecies;
 	}
-	public int getNumberOfSurveys() {
+	public long getNumberOfSurveys() {
 		return numberOfSurveys;
 	}
-	public void setNumberOfSurveys(int numberOfSurveys) {
+	public void setNumberOfSurveys(long numberOfSurveys) {
 		this.numberOfSurveys = numberOfSurveys;
 	}
+	
 	
 	
 }
